[1] TESTS - Test des associations du modèle Schedule

+++ src/test/java/models/TestSchedule.java
@@ -3,6 +3,8 @@ package models;
 import org.junit.jupiter.api.*;
 
 import javax.persistence.PersistenceException;
+import java.util.ArrayList;
+import java.util.Date;
 import java.util.List;
 
 import static java.util.Objects.nonNull;
@@ -14,19 +16,27 @@ import static org.junit.jupiter.api.Assertions.*;
 @TestMethodOrder(MethodOrderer.OrderAnnotation.class)
 public class TestSchedule implements TestModel {
 
-  private static int ID;
+  private static int ID_SESSION;
+  private static int ID_SERVER;
+  private static int ID_SCHEDULE;
   private static Schedule SCHEDULE;
   private static final String PROMOTION_TEST = "Promotion test";
   private static final String URL_TEST = "url.com";
   private static final String PROMOTION_UPDATE = "Promotion mise à jour";
   private static final String URL_UPDATE = "url.update.com";
+  public static final String REF_SERVER = "0132456789";
+  public static final String SESSION_NAME = "session";
+  public static final String SESSION_TEACHER = "prof";
+  public static final String SESSION_LOCATION = "location";
 
   @AfterAll
   public static void tearDown() {
-    SCHEDULE = Model.read(ID, Schedule.class);
-    if (nonNull(SCHEDULE)) {
-      SCHEDULE.delete();
-    }
+    Session session = Model.read(ID_SESSION, Session.class);
+    if (nonNull(session)) session.delete();
+    Server server = Model.read(ID_SERVER, Server.class);
+    if (nonNull(server)) server.delete();
+    SCHEDULE = Model.read(ID_SCHEDULE, Schedule.class);
+    if (nonNull(SCHEDULE)) SCHEDULE.delete();
   }
 
   @Test
@@ -34,9 +44,9 @@ public class TestSchedule implements TestModel {
   @Override
   public void testCreate() {
     SCHEDULE = new Schedule(PROMOTION_TEST, URL_TEST);
-    ID = SCHEDULE.create();
+    ID_SCHEDULE = SCHEDULE.create();
     List<Schedule> schedules = Model.readAll(Schedule.class);
-    assertTrue(schedules.stream().anyMatch(s -> s.getId() == ID));
+    assertTrue(schedules.stream().anyMatch(s -> s.getId() == ID_SCHEDULE));
   }
 
   @Test
@@ -59,7 +69,7 @@ public class TestSchedule implements TestModel {
   @Order(3)
   @Override
   public void testRead() {
-    Schedule schedule = Model.read(ID, Schedule.class);
+    Schedule schedule = Model.read(ID_SCHEDULE, Schedule.class);
     assertAll(
       () -> assertNotNull(schedule),
       () -> assertEquals(schedule.getId(), SCHEDULE.getId()),
@@ -70,6 +80,20 @@ public class TestSchedule implements TestModel {
 
   @Test
   @Order(4)
+  public void testAssociations() {
+    ID_SERVER = new Server(REF_SERVER, SCHEDULE).create();
+    ID_SESSION = new Session(SESSION_NAME, SESSION_TEACHER, SESSION_LOCATION, new Date(), new Date(), new Date(), SCHEDULE).create();
+    Schedule finalSchedule = Model.read(ID_SCHEDULE, Schedule.class);
+    assertAll(
+      () -> assertEquals(finalSchedule.getServers().size(), 1),
+      () -> assertEquals(new ArrayList<>(finalSchedule.getServers()).get(0).getId(), ID_SERVER),
+      () -> assertEquals(finalSchedule.getSessions().size(), 1),
+      () -> assertEquals(new ArrayList<>(finalSchedule.getSessions()).get(0).getId(), ID_SESSION)
+    );
+  }
+
+  @Test
+  @Order(5)
   public void testUpdate_promotion_null() {
     Schedule schedule = new Schedule(null, URL_UPDATE);
     schedule.setId(SCHEDULE.getId());
@@ -77,7 +101,7 @@ public class TestSchedule implements TestModel {
   }
 
   @Test
-  @Order(4)
+  @Order(6)
   public void testUpdate_url_null() {
     Schedule schedule = new Schedule(PROMOTION_UPDATE, null);
     schedule.setId(SCHEDULE.getId());
@@ -85,13 +109,13 @@ public class TestSchedule implements TestModel {
   }
 
   @Test
-  @Order(5)
+  @Order(7)
   @Override
   public void testUpdate() {
     SCHEDULE.setPromotion(PROMOTION_UPDATE);
     SCHEDULE.setUrl(URL_UPDATE);
     SCHEDULE.update();
-    SCHEDULE = Model.read(ID, Schedule.class);
+    SCHEDULE = Model.read(ID_SCHEDULE, Schedule.class);
     assertNotNull(SCHEDULE);
     assertAll(
       () -> assertEquals(SCHEDULE.getPromotion(), PROMOTION_UPDATE),
@@ -100,11 +124,21 @@ public class TestSchedule implements TestModel {
   }
 
   @Test
-  @Order(6)
+  @Order(8)
+  public void testDelete_whith_existing_associations() {
+    assertThrows(PersistenceException.class, () -> SCHEDULE.delete());
+  }
+
+  @Test
+  @Order(9)
   @Override
   public void testDelete() {
+    Session session = Model.read(ID_SESSION, Session.class);
+    if (nonNull(session)) session.delete();
+    Server server = Model.read(ID_SERVER, Server.class);
+    if (nonNull(server)) server.delete();
     SCHEDULE.delete();
-    Schedule schedule = Model.read(ID, Schedule.class);
+    Schedule schedule = Model.read(ID_SCHEDULE, Schedule.class);
     assertNull(schedule);
   }
 }
